+++ webapp/CHANGELOG.md
@@ -7,10 +7,9 @@
 
 ## [0.71.2](https://github.com/SocialGouv/carte-jeune-engage/compare/v0.71.1...v0.71.2) (2024-11-29)
 
-
 ### Bug Fixes
 
-* add missing migrations ([827a31e](https://github.com/SocialGouv/carte-jeune-engage/commit/827a31ece7352ae9ca7c057c9f42858e7a4dd089))
+- add missing migrations ([827a31e](https://github.com/SocialGouv/carte-jeune-engage/commit/827a31ece7352ae9ca7c057c9f42858e7a4dd089))
 
 ## [0.71.1](https://github.com/SocialGouv/carte-jeune-engage/compare/v0.71.0...v0.71.1) (2024-11-29)